+++ Adamantium/Adamantium.Imaging/Gif/GifDecoder.cs
@@ -94,7 +94,7 @@ namespace Adamantium.Imaging.Gif
             description.ArraySize = 1;
             description.Depth = 1;
 
-            var img = Image.New3D(gif.Descriptor.Width, gif.Descriptor.Height, gif.Frames.Count, new MipMapCount(1), SurfaceFormat.R8G8B8A8.UNorm);
+            var img = Image.New3D(gif.Descriptor.Width, gif.Descriptor.Height, (uint)gif.Frames.Count, new MipMapCount(1), SurfaceFormat.R8G8B8A8.UNorm);
             //var img = Image.New3D(gif.Descriptor.Width, gif.Descriptor.Height, 1, new MipMapCount(1), SurfaceFormat.R8G8B8A8.UNorm);
             //var img = Image.New3D(gif.Frames[45].Descriptor.Width, gif.Frames[45].Descriptor.Height, 1, new MipMapCount(1), SurfaceFormat.R8G8B8A8.UNorm);
             for (int i = 0; i < gif.Frames.Count; i++)
